Cleaned up what we expose via bindgen.
